+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Fixes
 
 - [#2219](https://gitlab.com/meltano/meltano/-/issues/2219) Don't fail on large (record) messages output by extractors (Singer taps) by increasing subprocess output buffer size from 64KB to 1MB.
+- [#2215](https://gitlab.com/meltano/meltano/-/issues/2215) Have `meltano invoke <plugin> --help` pass `--help` flag to plugin, instead of showing `meltano invoke` help message
 
 ### Breaks